requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.17.3
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.8.4
networkx==3.3
numpy==1.26.4
opencv-python==4.8.0.76
packaging==23.2
pandas==2.2.3
Pillow==9.3.0
protobuf==4.25.5
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
six==1.16.0
smmap==5.0.1
streamlit==1.25.0
sympy==1.13.3
tenacity==8.5.0
tokenizers==0.14.1
toml==0.10.2
torch==2.0.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.34.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
zipp==3.20.2
#brew install tesseract
#brew install tesseract-lang